Fix Debian trixie build break (#23)
Several fixes were needed to get this running on Debian trixie and
fixing Github Actions issues.

Apparently the members of FILE changed and the new way to initialise
FILE should be more robust, because it will blanket zero all members.

Creation of coverage report was broken, because it aborted as soon as
lcov reported an error. The reported errors are now ignored. The errors
are caused by macros from cppunit and the only other option would have
been to use another unit test framework.

valgrind on Github actions reported an issue at ping, which is used for
testing purposes. The issue is suppressed, because I did not want to
change test code too much.