clang-tidy fixes (#29)
The code is mostly used as an example how I would at minimum maintain a
C++ project and what tools and checks I use. Therefore the code should
be modernised sometimes.

Due to the lack of a nightly and the linter action only checking changed
files, a few findings have been accumulated since the switch to C++23
and LLVM 19.